% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww2stats-package.R
\docType{data}
\name{CasualtiesofWW2}
\alias{CasualtiesofWW2}
\title{World War 2 Casualties for each individual participating nation.}
\format{
A data frame with 56 rows and 4 variables:
\describe{
\item{Country}{Names of countries as factor}
\item{Military}{Number of casualties of military personnel}
\item{Civilian}{Number of casualties of civilians}
\item{Total}{Total casualties}

}
}
\source{
National WW2 Museum of New Orleans (https://www.nationalww2museum.org/students-teachers/student-resources/research-starters/research-starters-worldwide-deaths-world-war) Data accessed: 2024-11-01
}
\usage{
data(CasualtiesofWW2)
}
\description{
A dataset containing information about all particpators in WW2, including
civilian, and/or military, and/or total deaths depending on how each nation compiled
their casualties.
}
\examples{
 summary(CasualtiesofWW2)
}
\keyword{datasets}
