% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyr.R
\name{multiplyr}
\alias{multiplyr}
\title{A function for multiplying two numbers
A baby function to illustrate how to document stuff in R
it is simply a function to multiply two numbers}
\usage{
multiplyr(a, b)
}
\arguments{
\item{a}{a real number}

\item{b}{a real number}
}
\value{
The product of \code{a} and \code{b}
}
\description{
A function for multiplying two numbers
A baby function to illustrate how to document stuff in R
it is simply a function to multiply two numbers
}
\examples{
multiplyr(2, 5)
multiplyr(4, 8)
}
