% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww2stats-package.R
\docType{package}
\name{ww2stats-package}
\alias{ww2stats}
\alias{ww2stats-package}
\title{ww2stats: Data Package With Casualties For Participants}
\description{
Package is showing the casualties taken by participants in World War 2.
}
\author{
\strong{Maintainer}: Leonardo Paez-Goto \email{leonardopaezgoto@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
